// backend/src/modules/documents/document.types.ts
export enum DocumentCategory {
  LAND_DOCUMENTS = 'LAND_DOCUMENTS',
  PROJECT_DOCUMENTS = 'PROJECT_DOCUMENTS',
  AGREEMENTS = 'AGREEMENTS',
  INVESTOR_REQUESTS = 'INVESTOR_REQUESTS',
  SITE_REPORTS = 'SITE_REPORTS',
  OTHER = 'OTHER',
}

export enum DocumentVisibility {
  ADMIN_ONLY = 'ADMIN_ONLY',
  PROJECT_TEAM = 'PROJECT_TEAM',
  LANDOWNER = 'LANDOWNER',
  AUTHORIZED_INVESTOR = 'AUTHORIZED_INVESTOR',
}